Poursuite tests phase 2

+++ _lib/_pages_/concours/xmodules/admin/operations/phase_operations/phase_3.rb
@@ -11,7 +11,8 @@ class Operation
 
 # Méthode qui produit le fichier palmares.yaml qui contient les informations
 # sur les résultats pour le concours (cf. le manuel pour le détail)
-def consigne_resultats_in_file_palmares
+def consigne_resultats_in_file_palmares(options)
+  log("-> consigne_resultats_in_file_palmares(options=#{options.inspect})")
   # Structure de la donnée qui sera enregistrée dans palmares-<annee>.yaml
   dpalm = {
     infos:{
@@ -31,10 +32,10 @@ def consigne_resultats_in_file_palmares
   sans_dossier  = inscriptions.select { |dc| dc[:specs][0]    == '0' }
   non_conforme  = inscriptions.select { |dc| dc[:specs][0..1] == '12'}
 
-  hpalm[:infos][:nombre_inscriptions] = inscriptions.count
-  hpalm[:infos][:nombre_concurrents]  = concurrents.count
-  hpalm[:infos][:nombre_non_conforme] = non_conforme.count
-  hpalm[:infos][:nombre_sans_dossier] = sans_dossier.count
+  dpalm[:infos][:nombre_inscriptions] = inscriptions.count
+  dpalm[:infos][:nombre_concurrents]  = concurrents.count
+  dpalm[:infos][:nombre_non_conforme] = non_conforme.count
+  dpalm[:infos][:nombre_sans_dossier] = sans_dossier.count
 
   # Pour pouvoir procéder au calcul des notes et du classement
   require './_lib/_pages_/concours/xmodules/calculs/Dossier'
@@ -46,8 +47,13 @@ def consigne_resultats_in_file_palmares
   dpalm[:sans_dossier] = sans_dossier.collect { |dc| dc[:concurrent_id] }
 
   palmares_file_path = Dossier.palmares_file_path(ANNEE_CONCOURS_COURANTE)
-  File.delete(palmares_file_path) if File.exists?(palmares_file_path)
-  File.open(palmares_file_path,'wb'){|f| f.write(YAML.dump(dpalm))}
+
+  if options[:noop]
+    html.res << "Je dois créer le fichier #{palmares_file_path} avec la donnée palmarès suivante : #{dpalm.inspect}"
+  else
+    File.delete(palmares_file_path) if File.exists?(palmares_file_path)
+    File.open(palmares_file_path,'wb'){|f| f.write(YAML.dump(dpalm))}
+  end
 
 end #/ consigne_resultats_in_file_palmares
